requirements.txt
chromadb==1.0.15
groq==0.29.0
pandas==2.3.1
python-dotenv==1.1.1
rapidfuzz==3.13.0
semantic_router==0.1.9
streamlit==1.43.2
